Create cross_correlation.py
This code defines a function that performs a normalization and cross-correlated shift of a given spectrum according to an input out-of-transit spectrum. Once the code is run, the shift can be done with the following command: shift(normalized_spectra[i]) where i represents the spectrum index as part of a list of spectra.